Refactor folder structure and namespace of tests
