+++ composer.json
@@ -33,7 +33,7 @@
     "autoload-dev": {
         "psr-4": {
             "Francken\\Features\\": "tests/Features",
-            "Tests\\": "tests/"
+            "Francken\\Tests\\": "tests/Francken"
         }
     },
     "scripts": {
