+++ tests/Features/Repositories/IlluminateRepositoryFeature.php
@@ -3,6 +3,7 @@
 declare(strict_types=1);
 
 namespace Francken\Features\Repositories;
+
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
