+++ tests/Francken/Domain/Activities/ActivityTest.php
