+++ tests/Francken/Domain/Activities/Events/ActivityRescheduledTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Tests\Francken\Activities\Events;
 
-use Tests\SetupReconstitution;
+use Francken\Tests\SetupReconstitution;
 use Francken\Domain\Activities\ActivityId;
 use Francken\Domain\Activities\Schedule;
 use Francken\Domain\Activities\Events\ActivityRescheduled;
