+++ tests/Francken/Domain/Activities/ScheduleTest.php
