+++ tests/Francken/Domain/Committees/CommitteeTest.php
