+++ tests/Francken/Domain/Members/AddressTest.php
