+++ tests/Francken/Domain/Members/EmailTest.php
