+++ tests/Francken/Domain/Members/Registration/RegistrationRequestTest.php
