+++ tests/Francken/Domain/Posts/PostCategoryTest.php
