+++ tests/Francken/Domain/Posts/PostsTest.php
